# shared-contracts/openapi/Sacco.Api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "SACCO Platform API",
    "description": "Core banking API for Kenyan Deposit-Taking SACCOs. Pass the tenant slug in the X-Tenant header when not resolving by host. Tokens are issued by the in-process OIDC server at /connect/*.",
    "version": "v1"
  },
  "paths": {
...
    "/api/reporting/financial-position": {
      "get": {
        "tags": [
          "Regulatory reporting"
        ],
        "operationId": "GetFinancialPosition",
        "parameters": [
          {
            "name": "segment",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Segment"
            }
          },
          {
            "name": "asOf",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialPosition"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/income-statement": {
      "get": {
        "tags": [
          "Regulatory reporting"
        ],
        "operationId": "GetIncomeStatement",
        "parameters": [
          {
            "name": "segment",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Segment"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeStatement"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/capital-adequacy": {
      "get": {
        "tags": [
          "Regulatory reporting"
        ],
        "operationId": "GetCapitalAdequacy",
        "parameters": [
          {
            "name": "asOf",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapitalAdequacy"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/liquidity": {
      "get": {
        "tags": [
          "Regulatory reporting"
        ],
        "operationId": "GetLiquidity",
        "parameters": [
          {
            "name": "asOf",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LiquidityPosition"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/large-exposures": {
      "get": {
        "tags": [
          "Regulatory reporting"
        ],
        "operationId": "GetLargeExposures",
        "parameters": [
          {
            "name": "asOf",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LargeExposureReport"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/portfolio-quality": {
      "get": {
        "tags": [
          "Regulatory reporting"
        ],
        "operationId": "GetPortfolioQuality",
        "parameters": [
          {
            "name": "asOf",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioQualitySnapshot"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/preview": {
      "get": {
        "tags": [
          "Regulatory reporting"
        ],
        "operationId": "PreviewStatutoryReturn",
        "parameters": [
          {
            "name": "periodStart",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "periodEnd",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatutoryReturnPackage"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/statutory-returns": {
      "get": {
        "tags": [
          "Statutory returns"
        ],
        "operationId": "ListStatutoryReturns",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatutoryReturnSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Statutory returns"
        ],
        "operationId": "GenerateStatutoryReturn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatutoryReturnDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/statutory-returns/{id}": {
      "get": {
        "tags": [
          "Statutory returns"
        ],
        "operationId": "GetStatutoryReturn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatutoryReturnDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/statutory-returns/{id}/submit": {
      "post": {
        "tags": [
          "Statutory returns"
        ],
        "operationId": "SubmitStatutoryReturn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReturnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatutoryReturnSummary"
                }
              }
            }
          }
        }
      }
    },
    "/api/reporting/statutory-returns/{id}/withdraw": {
      "post": {
        "tags": [
          "Statutory returns"
        ],
        "operationId": "WithdrawStatutoryReturn",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatutoryReturnSummary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgingBucketSnapshot": {
        "required": [
          "name",
          "minDaysInArrears",
          "provisionRateBps",
          "loans",
          "outstanding",
          "provisionRequired"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "minDaysInArrears": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "provisionRateBps": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "loans": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "outstanding": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "provisionRequired": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "CapitalAdequacy": {
        "required": [
          "asOf",
          "shareCapital",
          "institutionalCapital",
          "coreCapital",
          "totalAssets",
          "totalDeposits",
          "ratios",
          "source"
        ],
        "type": "object",
        "properties": {
          "asOf": {
            "type": "string",
            "format": "date"
          },
          "shareCapital": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "institutionalCapital": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "coreCapital": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalAssets": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalDeposits": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "ratios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ratio"
            }
          },
          "source": {
            "type": "string"
          }
        }
      },
...
      "FinancialPosition": {
        "required": [
          "segment",
          "asOf",
          "assets",
          "liabilities",
          "equity",
          "accumulatedSurplus",
          "interSegmentNetted",
          "totalAssets",
          "totalLiabilitiesAndEquity",
          "balances"
        ],
        "type": "object",
        "properties": {
          "segment": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Segment"
              }
            ]
          },
          "asOf": {
            "type": "string",
            "format": "date"
          },
          "assets": {
            "$ref": "#/components/schemas/ReportSection"
          },
          "liabilities": {
            "$ref": "#/components/schemas/ReportSection"
          },
          "equity": {
            "$ref": "#/components/schemas/ReportSection"
          },
          "accumulatedSurplus": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "interSegmentNetted": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalAssets": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "totalLiabilitiesAndEquity": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "balances": {
            "type": "boolean"
          }
        }
      },
...
      "GenerateReturnRequest": {
        "required": [
          "periodStart",
          "periodEnd"
        ],
        "type": "object",
        "properties": {
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "periodEnd": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "IncomeStatement": {
        "required": [
          "segment",
          "from",
          "to",
          "income",
          "expenses",
          "surplus"
        ],
        "type": "object",
        "properties": {
          "segment": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Segment"
              }
            ]
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "income": {
            "$ref": "#/components/schemas/ReportSection"
          },
          "expenses": {
            "$ref": "#/components/schemas/ReportSection"
          },
          "surplus": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "LargeExposure": {
        "required": [
          "memberId",
          "loans",
          "outstanding",
          "shareOfCoreCapitalBps"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "loans": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "outstanding": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "shareOfCoreCapitalBps": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "LargeExposureReport": {
        "required": [
          "asOf",
          "coreCapital",
          "thresholdBps",
          "exposures",
          "source"
        ],
        "type": "object",
        "properties": {
          "asOf": {
            "type": "string",
            "format": "date"
          },
          "coreCapital": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "thresholdBps": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "exposures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LargeExposure"
            }
          },
          "source": {
            "type": "string"
          }
        }
      },
...
      "LiquidityPosition": {
        "required": [
          "asOf",
          "liquidAssets",
          "deposits",
          "shortTermLiabilities",
          "ratio",
          "source"
        ],
        "type": "object",
        "properties": {
          "asOf": {
            "type": "string",
            "format": "date"
          },
          "liquidAssets": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "deposits": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "shortTermLiabilities": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "ratio": {
            "$ref": "#/components/schemas/Ratio"
          },
          "source": {
            "type": "string"
          }
        }
      },
...
      "MemberOutstandingSnapshot": {
        "required": [
          "memberId",
          "loans",
          "outstanding"
        ],
        "type": "object",
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "loans": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "outstanding": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "PortfolioQualitySnapshot": {
        "required": [
          "asOf",
          "totalOutstanding",
          "nonPerformingOutstanding",
          "provisionRequired",
          "buckets",
          "byMember",
          "configSource"
        ],
        "type": "object",
        "properties": {
          "asOf": {
            "type": "string",
            "format": "date"
          },
          "totalOutstanding": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "nonPerformingOutstanding": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "provisionRequired": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "buckets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgingBucketSnapshot"
            }
          },
          "byMember": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberOutstandingSnapshot"
            }
          },
          "configSource": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "Ratio": {
        "required": [
          "name",
          "numerator",
          "denominator",
          "valueBps",
          "minimumBps",
          "compliant",
          "basis"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "numerator": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "denominator": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "valueBps": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "minimumBps": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "compliant": {
            "type": "boolean"
          },
          "basis": {
            "type": "string"
          }
        }
      },
...
      "ReasonRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
      "ReconciliationResult": {
        "required": [
          "isReconciled",
          "checks",
          "failures"
        ],
        "type": "object",
        "properties": {
          "isReconciled": {
            "type": "boolean"
          },
          "checks": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "failures": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ReportLine": {
        "required": [
          "code",
          "name",
          "amount"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
      "ReportSection": {
        "required": [
          "title",
          "lines",
          "total"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReportLine"
            }
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "ReturnStatus": {
        "enum": [
          "Generated",
          "Submitted",
          "Withdrawn"
        ]
      },
...
      "Segment": {
        "enum": [
          "Fosa",
          "Bosa"
        ]
      },
...
      "StatutoryReturnDetail": {
        "required": [
          "summary",
          "package"
        ],
        "type": "object",
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/StatutoryReturnSummary"
          },
          "package": {
            "$ref": "#/components/schemas/StatutoryReturnPackage"
          }
        }
      },
      "StatutoryReturnPackage": {
        "required": [
          "periodStart",
          "periodEnd",
          "generatedAt",
          "consolidatedPosition",
          "fosaPosition",
          "bosaPosition",
          "consolidatedIncome",
          "fosaIncome",
          "bosaIncome",
          "capitalAdequacy",
          "liquidity",
          "portfolioQuality",
          "largeExposures",
          "sdgfContribution",
          "reconciliation",
          "openItems"
        ],
        "type": "object",
        "properties": {
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "periodEnd": {
            "type": "string",
            "format": "date"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "consolidatedPosition": {
            "$ref": "#/components/schemas/FinancialPosition"
          },
          "fosaPosition": {
            "$ref": "#/components/schemas/FinancialPosition"
          },
          "bosaPosition": {
            "$ref": "#/components/schemas/FinancialPosition"
          },
          "consolidatedIncome": {
            "$ref": "#/components/schemas/IncomeStatement"
          },
          "fosaIncome": {
            "$ref": "#/components/schemas/IncomeStatement"
          },
          "bosaIncome": {
            "$ref": "#/components/schemas/IncomeStatement"
          },
          "capitalAdequacy": {
            "$ref": "#/components/schemas/CapitalAdequacy"
          },
          "liquidity": {
            "$ref": "#/components/schemas/LiquidityPosition"
          },
          "portfolioQuality": {
            "$ref": "#/components/schemas/PortfolioQualitySnapshot"
          },
          "largeExposures": {
            "$ref": "#/components/schemas/LargeExposureReport"
          },
          "sdgfContribution": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "reconciliation": {
            "$ref": "#/components/schemas/ReconciliationResult"
          },
          "openItems": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "StatutoryReturnSummary": {
        "required": [
          "id",
          "periodStart",
          "periodEnd",
          "status",
          "generatedByUserId",
          "generatedAt",
          "submittedByUserId",
          "submittedAt",
          "submissionReference",
          "isReconciled",
          "reconciliationNotes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "periodStart": {
            "type": "string",
            "format": "date"
          },
          "periodEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "$ref": "#/components/schemas/ReturnStatus"
          },
          "generatedByUserId": {
            "type": "string",
            "format": "uuid"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "submittedByUserId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "submittedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "submissionReference": {
            "type": [
              "null",
              "string"
            ]
          },
          "isReconciled": {
            "type": "boolean"
          },
          "reconciliationNotes": {
            "type": "string"
          }
        }
      },
...
      "SubmitReturnRequest": {
        "required": [
          "submissionReference"
        ],
        "type": "object",
        "properties": {
          "submissionReference": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}